refactor(analysis-types): move topics serde test
